Move SetType implementation to a new types.py file that should be the default location for any new field type definitions
