+++ cc_dynamodb3/models.py
@@ -8,8 +8,6 @@ import uuid
 
 from schematics.models import Model
 from schematics import types as fields
-from schematics.types.compound import ListType
-
 
 from botocore.exceptions import ClientError
 
@@ -17,29 +15,6 @@ from . import exceptions
 from .config import get_config
 from .log import log_data
 from .table import get_table, query_table
-
-
-class SetType(ListType):
-
-    def _force_set(self, value):
-        if value is None or value == set():
-            return set()
-
-        try:
-            if isinstance(value, basestring):
-                raise TypeError()
-
-            if isinstance(value, dict):
-                return set(value[unicode(k)] for k in sorted(map(int, value.keys())))
-
-            return set(value)
-        except TypeError:
-            return set(value)
-
-    def to_native(self, value, context=None):
-        items = self._force_set(value)
-
-        return set(self.field.to_native(item, context) for item in items)
 
 
 class DynamoDBModel(Model):
